Renamings in plugin directories and targets
Removed the Plugin suffix from plugin subdirectories,

Consistent naming for all plugins libraries: braynsFoo instead of
braynsFooPlugin or FooPlugin.

Adding VRPN plugin target to Brayns-all
