+++ ospray_modules/CMakeLists.txt
@@ -10,7 +10,7 @@ option(BRAYNS_OPTIX_ENABLED "Activate OptiX module" OFF)
 if(BRAYNS_OPTIX_ENABLED)
   add_subdirectory(optix)
   if(TARGET ospray_module_optix)
-    add_dependencies(braynsOSPRayEnginePlugin ospray_module_optix)
+    add_dependencies(braynsOSPRayEngine ospray_module_optix)
   endif()
 endif()
 
@@ -19,7 +19,7 @@ option(BRAYNS_OPENDECK_ENABLED "Activate OpenDeck module" OFF)
 if(BRAYNS_OPENDECK_ENABLED)
   add_subdirectory(opendeck)
   if(TARGET ospray_module_opendeck)
-    add_dependencies(braynsOSPRayEnginePlugin ospray_module_opendeck)
+    add_dependencies(braynsOSPRayEngine ospray_module_opendeck)
   endif()
 endif()
 
@@ -28,6 +28,6 @@ option(BRAYNS_STEREOSCOPY_ENABLED "Activate Stereoscopy module" OFF)
 if(BRAYNS_STEREOSCOPY_ENABLED)
   add_subdirectory(stereoscopy)
   if(TARGET ospray_module_stereoscopy)
-    add_dependencies(braynsOSPRayEnginePlugin ospray_module_stereoscopy)
+    add_dependencies(braynsOSPRayEngine ospray_module_stereoscopy)
   endif()
 endif()
