+++ plugins/Deflect/DeflectPlugin.cpp
