+++ plugins/Example/CMakeLists.txt
@@ -4,12 +4,12 @@
 # This file is part of Brayns <https://github.com/BlueBrain/Brayns>
 
 cmake_minimum_required(VERSION 3.1 FATAL_ERROR)
-project(ExamplePlugin VERSION 1.0.0)
+project(ExampleBraynsPlugin VERSION 1.0.0)
 
 find_package(Brayns REQUIRED)
 
-add_library(examplePlugin SHARED
-  ExamplePlugin.cpp
-  ExamplePlugin.h)
+add_library(exampleBraynsPlugin SHARED
+  Example.cpp
+  Example.h)
 
-target_link_libraries(examplePlugin braynsPluginAPI)
+target_link_libraries(exampleBraynsPlugin braynsPluginAPI)
