+++ plugins/Example/Plugin.cpp
