+++ plugins/Example/Plugin.h
