+++ plugins/Rockets/ImageGenerator.cpp
