+++ plugins/Rockets/ImageGenerator.h
