+++ plugins/Rockets/SnapshotTask.h
