+++ plugins/Rockets/staticjson/VERSION
