+++ plugins/Rockets/staticjson/enum.hpp
