+++ plugins/Rockets/staticjson/forward_declarations.hpp
