+++ plugins/Rockets/staticjson/staticjson.cpp
